requirements.txt
djangorestframework
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
Collectfast==0.5.2
constantly==15.1.0
daphne==1.3.0
Django==1.11.7
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.3
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
Markdown==2.6.10
msgpack-python==0.4.8
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
slackclient==1.1.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
websocket-client==0.44.0
zope.interface==4.4.3
